Request callbacks all recieve a single argument.
The done, fail and always methods of AWSRequest all
receive a single argument when called.  This argument
is the AWSResponse object (previously called the
context object).

The AWSRequest callback methods (done, fail and always)
now receive a single param when called.  This is an
AWSResponse object.

  var req = ddb.listTables();

  req.done(function(resp) {
    console.log(resp.data);
  });

  req.fail(function(resp) {
    console.log(resp.error);
  });

  req.always(function(resp) {
    if (resp.error) {
      console.log(resp.error);
    } else {
      console.log(resp.data);
    }
  });

You can now specify what object to bind callbacks to when
registering a callback via done, fail and always.

  var req = ddb.listTables();
  req.done(tableDisplayWidget.refresh, { bind: tableDisplayWidget });