Add global config object AWS.configuration
AWS.configuration can be set to an AWS.Config object or an object literal,
e.g.,

    AWS.configuration = {
      accessKeyId: 'accessKey',
      secretAccessKey: 'secret',
      region: 'us-west-1'
    };

This configuration is used as the set of defaults for all Service
instantiation. This means that the config parameter in Service is now
optional. If not provided, the base configuration will be merged in.

The configuration object is always copied when merged with new options
so that the global configuration object reference is never directly passed
to a Service. The following contract always holds:

    var config = new AWS.Config(...);
    var service = Service.new(config);
    service.config == config;            // FALSE!
    service.config != config;            // TRUE!
    service.config != AWS.configuration; // TRUE!