Build Content-Length after service interfaces
Adds a new 'afterBuild' hook event and listener named
AWS.EventListeners.Core.SET_CONTENT_LENGTH to build Content-Length
on headers after services build the request object. This ensures
that content length is set consistently for all input types regardless
of the service interface.

Note: afterBuild is a private event and is reserved for internal use only.

Related to (but does not completely fix) #30