Do not retry a streaming operation if it started
This change disables retry on Request.createReadStream()
once the stream event listeners have been registered. Errors
propagate up to the stream (and request) and are not retryable.

This fixes issues where a stream is in the middle of a request
and a network error occurs, forcing a retry of the request and
corrupting the output stream.