Add support for loading clients by API version
To load a client by a given API version, pass the "apiVersion"
setting to the constructor:

    var ec2 = new AWS.EC2({apiVersion: '2013-02-01'});

A version date can also be "fuzzy". If there is no exact matching
API version, the SDK will look for the latest matching date. For
instance, passing '2013-12-12' will still use the above EC2 API
version of '2013-02-01' (and any subsequent version until the
end of the year).

You can also pass the value 'latest' to grab the latest API version
available for the service:

    var ec2 = new AWS.EC2({apiVersion: 'latest'});

Finally, you can pass 'apiConfig' to a service to create one-off
clients using a loaded API configuration:

    var myConfig = require('myApiConfig');
    var myClient = new AWS.Client({apiConfig: myConfig});