Merge Client into Service class, drop .client property
Also renames AWS.Request.client to AWS.Request.service to be
consistent with the new service concept.

UPGRADING NOTES:

1. AWS.Request.client renamed to AWS.Request.service:

This is a backward incompatible change that is scoped to the
AWS.Request class. If you use this object, change:

`request.client.params` to `request.service.params`

Or any other property access on `request.client`.

2. Client and client properties deprecated:

SERVICE.Client and service.client properties remain for backward
compatibility, but are deprecated and no longer documented. They
may be removed in a 1.0 release. To migrate these properties, change:

`new AWS.S3.Client(...)` to `new AWS.S3(...)`

and/or:

`s3.client.listObjects(...)` to `s3.listObjects(...)`