Fix conflict in Expires param for S3.putObject()
A change was made the Signers.S3 module to use the Expires header
if present as the signed date, as per support for pre-signed URLs
(getSignedUrl). This field conflicts with an Expires parameter which
can be passed into a putObject operation. This fix renames the
internal header that is checked to differentiate the putObject Expires
header from the getSignedUrl Expires parameter.

Fixes #120