Implicitly convert int/float params in strings
This change allows for integer and float parameters that are
represented as strings ("1.23", or "123") to be valid inputs and
implicitly converted to numbers when sent over the wire.

This became an issue in commit 083f08e due to a change of PartNumber
in the S3 API from a string to its proper integer type, which was a
backward incompatible API bug fix. This commit makes the previous
change backward compatible by implicitly converting types.

Fixes #211