Fix error events not firing in createReadStream()
The addition of support for HTTP progress events caused a
re-ordering of event listeners making it so readable/data events
would never get attached to the HTTP stream when a request was made
with createReadStream(). Because no events were registered, Node.js's
HTTP library emits an 'aborted' event instead of 'end', which the
SDK was not prepared to handle (and tests did not mock).

The fix for this issue conditionally sets the streamable flag only
when a successful request is made (2xx status code), allowing data
events to be listened for in the cases of 3xx, 4xx, or 5xx responses.

Fixes #214