Breaking change: convert binary and timestamp types on JSON output
This change modifies the default behavior of types returned from
JSON based services. By default, AWS.Config.convertResponseTypes
is set to true, which does type conversion on binary (base64) and
timestamp types returned by services. If performance of this parsing
is an issue, it is possible to return the raw JSON response by setting
config.convertResponseTypes to false.