Signal Initial Validity
A small change to resolve #1243 by exposing the role_arn value on the attribute roleArn after the initial validation of the settings.  Signal that the credentials have passed initial validation checks (i.e. the role_arn and source_profiles values were set and the source_profile was successfully loaded for use in invoking STS).

Alter unit tests to validate at least once that the roleArn attribute is created and available in a valid configuration unit test.