Add a `wrapNumbers` option to the DynamoDB Document Client
This option instructs the Document Client not to convert
NumberAttributeValues or members of NumberSetAttributeValues
to JavaScript numbers, but to instead use a wrapper object
that uses the string value retrieved from DynamoDB as its
source of truth. This allows values that cannot be precisely
stored in an IEEE 754 double to be round-tripped without
losing precision.