Support shared ini file mfa_serial with fn to fetch token code
`tokenCodeFn` will be called if mfa_serial is found in the active profile of the
shared ini file. This function will be called with the value of `mfa_serial` and
should invoke the callback with the MFA token or an error. This token will be
provided to the STS assume role call.