Fix link formatting in readme
Markdown is not rendered inside of HTML blocks, so having a `[link](...)` inside of a `<p></p>` block doesn't render as intended. I've removed the paragraph, because it has no effect here on Github. I'm not aware of any other documentation websites that might be displaying this readme, but if one exists I can update my PR to put the paragraph back and use HTML to format the link.