Create ChainableTemporaryCredentials and standardize coalescing of concurrent refresh calls (#2175)
* feat: new ChainableTemporaryCredentials class

ChainableTemporaryCredentials facilitate arbitrary AssumeRole paths.

TemporaryCredentials bypasses the masterCredentials if it was an instance
of TemporaryCredentials. This precluded using TemporaryCredentials to
traverse roles. For example, User A -> Role A -> Role B. In this case,
Role B cannot be refreshed using User A's credentials.

* refactor: unify coalescing refresh calls for all async refreshable credentials

Credentials with refresh methods that are synchronous remain as they were
to avoid potentially breaking users who expected the credentials to load
synchronously.

Credentials with an async refresh method now coalesce refresh calls in
a common way by implementing a load function which is responsible for
loading the credential properties and updating the Credentials instance.

All Credentials which use coalesceRefresh run the same coalesce test.

RemoteCredentials.refresh now respects the constructor options when
an instance is refreshed. Previously, all instances shared a static
refreshQueue, which meant the options of the first instance in the queue
were used for the singular refresh call that all instances would share.

SharedIniFileCredentials constructor now accepts a callback option, since
its refresh method is called implicitly during construction. The
constructor no longer throws an error when the HOME dir does not exist
because the refresh method now ensures that any errors are passed to
callback instead of potentially throwing synchronously. The caller
can pass AWS.util.fn.callback as the callback option to make the
constructor throw, but the default callback is AWS.util.fn.noop because
the default CredentialProvider expects the constructor to swallow any
async errors that may occur.

All credentials tests that test async behavior are now written to work
with async code.

Updated eslintrc to warn about missing semicolons.

* update ChainableTemporaryCredentials docs and support MFA

* fix: correct ChainableTemporaryCredentials constructor docs

* fix: prune refreshCallbacks assignment in derived Credentials classes