Allow overwriting STS config in ChainableTemporaryCredentials
STS uses the global endpoint per default which does not work for
"opt-in" regions. The `ChainableTemporaryCredentials` wrapper for
fetching credentials via `AssumeRole` or `GetSessionToken` does not
allow overriding the endpoint or set a different region.

This change will introduce another optional constructor parameter for
specifying the STS client configuration and forwarding it to the client
constructor.

In addition I remove the incorrect typing for the second
`masterCredentials` constructor parameter. The parameter needs to be
passed as a nested value in `options`.

issue: https://github.com/aws/aws-sdk-js/issues/2673