Allow shortcutting in customBackoff based on error
Adds the error as a parameter to the `customBackoff` function. This should be a non-breaking change for all clients.

If the result of the custom backoff is negative, then use that as a signal to shortcut remaining retries. This would make explicit a previously undefined behavior if clients returned a negative value from `customBackoff`.