Simplify consumption of client generation tools
This is a backwards compatible change to make the client generator,
typings generator and minimize json easier to consume for building
clients not included in the SDK.

Changes are:

* Paths are not hard-coded anymore but instead a `basePath` can be
provided which specifies the path to the sdk.
* Make `translate-api`, `client-creator.js` and `ts-generator.js`
callable as script and accept an optional `basePath` parameter
* Make `translate-api` optionally consumable as class. `Translator` is
already consumable but having a generic reading and writing of the
minimized file is beneficial as well.
* It removes side-effects from constructors to avoid unexpected changes
on initialization.