Fix nextPage typing in requests
The fixes the return type of the `nextPage()` method. Before, it would
return `Request<D, E>|void` which is hard to work with, because you
couldn't runtime test for `void` (everything can be void). This change
splits up the definition into two overloads: one without a callback
provided that returns `Request<D, E> | null` and a second for the
callback version that always returns `void`. This also matches the
documentation in the implementation now.