Re-export ConfigBase to make client-generation backwards compatible
https://github.com/aws/aws-sdk-js/commit/9bd7fdbda9929d4f259996d468de80e210a4ce78
changed the interface for generated clients. That works great as long as
the client is generated with the same aws-sdk-js version as it's
consumed which is generally the case since the whole aws-sdk is shipped
as one package. But in case of a private aws-sdk client it will break.
When the sdk version used for generation is >2.740.0 the consumer needs
to have a newer version as well. And the other way round.

This change will at least keep the old interface. A client generated
with <2.740.0 will be sill consumable with a newer aws-sdk but not the
other way round.