Fix typings for AWSError
AWSError is not a class. It's based on Error but with additional properties on top:

https://github.com/aws/aws-sdk-js/blob/master/lib/util.js#L570-L603

By wrongly declaring a class statements like `if (error instanceof AWSError)` would compile but during runtime `AWSError` is `undefined` since it's just a type. Fixing the types prevents this issue. In addition many properties are optional. Declaring them as required can cause type errors and makes mocking more difficult.