Fix dual-callback case (#3537)
This fixes a bug where using the callback version of createPresignedPost
without proper AWS credentials configured would result in an
un-catchable error and crash the node process.

It is not possible to write a test to capture this error due to the
nature of the error.  The callback is called, and only after that does
the crash occur.