fix: update redirect logic to also update path
When a 307 redirect comes in, it contains the new, full URL for the location change.
The current logic respects this for the `Host` header that gets sent, but it does not
then update the path, leading to possibly invalid redirects. The new logic pulls the
path out of the endpoint and updates the outgoing request.