Sqs AwsJson migration (#4105)
* Add awsQueryCompatible trait support to service

When awsQueryCompatible trait is found, it's made available as an API property. When this property is found,
the error code is returned by looking up the mapping. This is a pre-requisite for migrating services from
AWSQuery wire protocol to AWSJson.

* Add additional tests; review comments

* Add awsQueryCompatible trait support to service

When awsQueryCompatible trait is found, it's made available as an API property. When this property is found,
the error code is returned by looking up the mapping. This is a pre-requisite for migrating services from
AWSQuery wire protocol to AWSJson.

* Add additional tests; review comments

* Review comments

* Use MockServiceFromApi; Move errorCodeMapping logic to even)listeners

* Update on.error function

* Fix type

* Remove reimplementation of mapping function

* Use add instead of addAsync

Co-authored-by: AllanZhengYP <zheallan@amazon.com>

* Add null check to error listener

* Use simple service null check

* Fix add 'Error' implementation

* Add changelog

Co-authored-by: AllanZhengYP <zheallan@amazon.com>