AWSQuery compatible error code (#4221)
* Add awsQueryCompatible trait support to service

When awsQueryCompatible trait is found, it's made available as an API property. When this property is found,
the error code is returned by looking up the mapping. This is a pre-requisite for migrating services from
AWSQuery wire protocol to AWSJson.

* Minor edit

* Add additional tests; review comments

* Minor edits

* Sqs aws json (#2)

* Add awsQueryCompatible trait support to service

When awsQueryCompatible trait is found, it's made available as an API property. When this property is found,
the error code is returned by looking up the mapping. This is a pre-requisite for migrating services from
AWSQuery wire protocol to AWSJson.

* Minor edit

* Add additional tests; review comments

* Minor edits

* Review comments

* Use MockServiceFromApi; Move errorCodeMapping logic to even)listeners

* Update on.error function

* Fix type

* Remove reimplementation of mapping function

* Use add instead of addAsync

Co-authored-by: AllanZhengYP <zheallan@amazon.com>

* Fix lint errors

* Add null check to error listener

* Use simple service null check

* Fix add 'Error' implementation

* Add changelog

* Revert "Sqs AwsJson migration (#4105)"

This reverts commit 2a87ec1b21ffc9b705d10fa8b3e5128f451b926b.

* Add AWSQuery error code compatibility to service

When  header value is found, then error code is populated from the header
If there is no header found, then the default error code is returned.

Co-authored-by: AllanZhengYP <zheallan@amazon.com>