Update ts-generator.js to stop typescript from collapsing strings with
This commit fixes an issue that's the current TS generator causes on the AWS-SDK.

As of today, when an API has an argument that is an enum like, this enum gets converted to a string literal that's then merged with `string`. ie: `'Foo' | 'Bar' | string`

While this works mostly okay, it comes with a major downside: Typescript loses reference to the original inputs and collapses everything into `string`, since `string` superseeds both `foo` and `bar`. This causes an issue that, if you're then using either of these types to type your function inputs, intellisense cannot provide type-ahead hints to it.

This PR fixes that behaviour by merging in `string & {}` instead of just `string`. By using it, we keep it fully backwards compatible and we solve the issue described above

--

Try it by yourself:

https://www.typescriptlang.org/play?#code/C4TwDgpgBAkgdmArsKBeKByAZgexxqAH0wCMBDAJwOIwFsIALDAbgChWtE4BjYASxxwoAYUQUKEOMABCjMgDcBYgBSUA5gGcAXLATIiUDcAp84agJQBvAL6tR4yTLmKcKjGQ0ATDObtiJUrIMCkoUyth4PuwcXLwCQgByEADuQSGuYeraukgoxMpGJmZQAGRQNuZWtkmpzqHh5FS+NWkubsAMfBoAtF3dZAA2GjjdOADWZCA+QA